requirements.txt
# Core libraries
flask          # frontend service
fastapi        # backend services (router, OCR, preprocessing, etc.)
uvicorn        # run FastAPI apps
requests       # HTTP calls between services
pydantic       # request/response validation in services
aiohttp        # async HTTP client for router service

# OCR and NLP libraries
opencv-python  # image preprocessing
numpy          # array ops (needed by OpenCV, ML models)
scikit-learn   # optional, for lang-detect / ML classifiers
langdetect     # quick language detection baseline
torch          # deep learning OCR models
transformers   # pretrained OCR/Lang models (HuggingFace)
pillow         # image manipulation

# Data handling
pandas         # tabular data / excel if needed
openpyxl       # .xlsx handling
python-docx    # .docx if still needed
pypdf2         # PDF parsing
pdf2image      # PDF to image conversion for preprocessing

# Object Detection
ultralytics        # YOLOv8/v11 for signature, stamp, QR detection
supervision        # optional: advanced visualization/tracking

# Utilities
python-multipart   # file uploads in FastAPI
pytest             # testing
